test: restructure private optimizer ITs into per-rule modules + README
Address amahussein's overall review on #14902:

1. Structure: split private_optimizer_test.py into flat per-rule modules
   (private_optimizer_{agg_pushdown,decompose_stddev,subquery_shared_scan,
   skewed_bhj_join}_test.py) sharing private_optimizer_common.py. The common
   module exposes private_optimizer_conf(*rule_confs, extra_conf=None) and the
   collect_and_plan / assert_rule_fires "teeth" (off-rule CPU baseline vs
   on-rule GPU run + per-rule plan marker). Flat layout, matching the repo (no
   nested pytest test dirs). Old single file removed.

2. Documentation: private_optimizer_README.md covers where to add a rule test,
   choosing the gate conf, making a rule reachable, the off-baseline +
   plan-marker pattern, when an aggregate-only check is acceptable, the
   compatibility mechanism, and local validation.

3. Compatibility: shared require_private_optimizer guard skips below Spark 3.3.0
   (the private plugin build floor: 330..411 + 400db173). Within that matrix all
   four rules apply on every runtime including Databricks, so no per-runtime
   skips are invented. The skew rule's internal isOptimizeSkewBHJSupported
   kill-switch is backstopped by the plan-marker assertion (fails loudly if a
   runtime drops support) rather than a guessed version.

Skew module keeps the GLOBAL-aggregate validation (a GROUP BY on the skew key
hits a known wrong-results defect).

Validated locally with run_pyspark_from_build.sh -m private_optimizer:
4 passed, 39895 deselected (collected from the 4 new modules).

### Review notes
Structural refactor; helper bodies are byte-identical to the prior
_collect_and_plan/_assert_rule_fires, the only test-body change is the
private_optimizer_conf call form (rule conf as positional arg, query-shaping
conf via extra_conf) which is value-equivalent since the dicts share no keys.
nt-code-review's multi-agent pipeline could not run here (no Agent dispatch in
this context); equivalence was verified by diffing each test body and the
helpers against HEAD.
