+++ integration_tests/pytest.ini
@@ -32,7 +32,7 @@ markers =
     fuzz_test: Mark fuzz tests
     iceberg: Mark a test that requires Iceberg has been configured, skipping if tests are not configured for Iceberg
     delta_lake: Mark a test that requires Delta Lake has been configured, skipping if tests are not configured for Delta Lake
-    private_optimizer: Mark a test that exercises a private optimizer rule (organizational; see private_optimizer_test.py for the covered rules).
+    private_optimizer: Mark a test that exercises a private optimizer rule (organizational; see private_optimizer_README.md for the covered rules and how to add more).
     allow_non_gpu_delta_write_if(bool, reason): Allow non-GPU writes to Delta Lake tables when the condition is met
     regexp: Mark a test that tests regular expressions on the GPU (only works when UTF-8 is enabled)
     large_data_test: Mark tests with large data
